// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/admin-ai-prompt-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface AdminAiPromptHttpResponse {
    createdOnUtc?: string;
    id?: string;
    isActive?: boolean;
    key?: string;
    locale?: string;
    promptText?: string;
    updatedOnUtc?: string | null;
    version?: number;
}