// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/admin-ai-prompt-scenario-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AdminAiPromptHttpResponse } from './admin-ai-prompt-http-response';

export interface AdminAiPromptScenarioHttpResponse {
    inheritedPromptText?: string;
    inheritedSource?: string;
    key?: string;
    locale?: string;
    promptText?: string;
    responseFormatJson?: string;
    source?: string;
    sourceLocale?: string;
    template?: AdminAiPromptHttpResponse | null;
    variables?: Array<string>;
}